add support for embeddings, tool calls, and streaming
