+++ js/examples/next-openai-telemetry-app/app/api/text/route.ts
@@ -1,9 +1,81 @@
 import { openai } from "@ai-sdk/openai";
-import { generateText, tool } from "ai";
+import {
+  embed,
+  embedMany,
+  generateObject,
+  generateText,
+  streamObject,
+  streamText,
+  tool,
+} from "ai";
 import { z } from "zod";
 
 export async function POST(req: Request) {
   const { prompt } = await req.json();
+  const textStream = await streamText({
+    model: openai("gpt-3.5-turbo"),
+    maxTokens: 100,
+    messages: [
+      { content: "hello, you are a great bot", role: "system" },
+      { content: prompt, role: "user" },
+    ],
+
+    experimental_telemetry: {
+      isEnabled: true,
+      metadata: { example: "value" },
+    },
+  });
+
+  await embed({
+    model: openai.embedding("text-embedding-ada-002"),
+    value: "hello, you are a great bot",
+    experimental_telemetry: {
+      isEnabled: true,
+      metadata: { example: "value" },
+    },
+  });
+
+  await embedMany({
+    model: openai.embedding("text-embedding-ada-002"),
+    values: ["hello, you are a great bot", prompt],
+    experimental_telemetry: {
+      isEnabled: true,
+      metadata: { example: "value" },
+    },
+  });
+
+  await generateObject({
+    model: openai("gpt-3.5-turbo"),
+    maxTokens: 100,
+    messages: [
+      { content: "hello, you are a great bot", role: "system" },
+      { content: "please create me 3 objects like this", role: "user" },
+    ],
+    schema: z.object({
+      description: z.string().describe("The generated text"),
+      number: z.number().describe("A generated number"),
+    }),
+    experimental_telemetry: {
+      isEnabled: true,
+      metadata: { example: "value" },
+    },
+  });
+  const objectStream = await streamObject({
+    model: openai("gpt-3.5-turbo"),
+    maxTokens: 100,
+    messages: [
+      { content: "hello, you are a great bot", role: "system" },
+      { content: "please create me 3 objects like this", role: "user" },
+    ],
+    schema: z.object({
+      description: z.string().describe("The generated text"),
+      number: z.number().describe("A generated number"),
+    }),
+    experimental_telemetry: {
+      isEnabled: true,
+      metadata: { example: "value" },
+    },
+  });
 
   const { text, toolResults } = await generateText({
     model: openai("gpt-3.5-turbo"),
@@ -30,7 +102,18 @@ export async function POST(req: Request) {
     },
   });
 
-  return new Response(JSON.stringify({ text, toolResults }), {
-    headers: { "Content-Type": "application/json" },
-  });
+  const textStreamResponse = await textStream.toTextStreamResponse().text();
+  const objectStreamResponse = await objectStream.toTextStreamResponse().text();
+
+  return new Response(
+    JSON.stringify({
+      text,
+      toolResults,
+      textStreamResponse,
+      objectStreamResponse,
+    }),
+    {
+      headers: { "Content-Type": "application/json" },
+    },
+  );
 }
