+++ js/packages/openinference-vercel-ai-sdk-span-processor/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./OpenInferenceSpanExporter";
