+++ js/packages/openinference-vercel-ai-sdk-span-processor/src/vercel/constants.ts
@@ -17,10 +17,19 @@ import { OpenInferenceSemanticConvention } from "./types";
  * @example ai.generateText.doGenerate my-chat-call
  */
 export const VercelSDKFunctionNameToSpanKindMap = new Map([
-  ["doGenerate", OpenInferenceSpanKind.LLM],
-  ["doStream", OpenInferenceSpanKind.LLM],
-  ["doEmbed", OpenInferenceSpanKind.EMBEDDING],
-  ["toolCall", OpenInferenceSpanKind.TOOL],
+  ["ai.generateText", OpenInferenceSpanKind.CHAIN],
+  ["ai.generateText.doGenerate", OpenInferenceSpanKind.LLM],
+  ["ai.generateObject", OpenInferenceSpanKind.CHAIN],
+  ["ai.generateObject.doGenerate", OpenInferenceSpanKind.LLM],
+  ["ai.streamText", OpenInferenceSpanKind.CHAIN],
+  ["ai.streamText.doStream", OpenInferenceSpanKind.LLM],
+  ["ai.streamObject", OpenInferenceSpanKind.CHAIN],
+  ["ai.streamObject.doStream", OpenInferenceSpanKind.LLM],
+  ["ai.embed", OpenInferenceSpanKind.CHAIN],
+  ["ai.embed.doEmbed", OpenInferenceSpanKind.EMBEDDING],
+  ["ai.embedMany", OpenInferenceSpanKind.CHAIN],
+  ["ai.embedMany.doEmbed", OpenInferenceSpanKind.EMBEDDING],
+  ["ai.toolCall", OpenInferenceSpanKind.TOOL],
 ]);
 
 export const VercelSemConvToOISemConvMap: Record<
